Make zunit init idempotent
Currently, running `zunit init` fails if either `.zunit.yml` or the
directory `tests` exists within the repository, meaning the default
`.travis.yml` cannot be generated. This commit checks for the existence
of `.zunit.yml`, `tests` and `.travis.yml` as before, but instead of
exiting with an error, raises a warning and continues.

Fixes #88