Ensure assertion value is double quoted
Internally, the `assert` helper does not quote the value to be asserted
against when cascading the call to the relevant assertion function,
resulting in an implicit pass if the value is empty, since the
comparisons are shifted forward a step. Double quoting the vaue when
cascading prevents these false-positives from occurring.

- [ ] Write tests to capture these false-positives

Fix #84