helpers.zsh/run: Allow running of complex commands containing ;, |, etc.
Basically, instead of:

     output=$("${cmd[@]}" 2>&1)

it is done:

     eval "output=$( ${cmd[@]} 2>&1)"

because this allows the command to include e.g. redirections or command
separators like ; or ||, etc., or pipes, etc.

Actually, the eval looks more terrifying:

  IFS=$'\n' eval "output=\$( function run {
        ${cmd[@]/(#m)*/${${${${${(M)MATCH:#(${(j:|:)~dont_quote})}:+$MATCH}}:-\"$MATCH\"}}} 2>&1 }; run )";

But the substitution is (as always..) simple: it quotes the command's arguments
by using double quotes, so that when eval sees it, it will not skip empty
strings (and also that it will not split strings). The dont_quote array holds
things that are not to be quoted – currently the semicolon and redirections
('[[:digit:]]>&[[:digit:]]). The run function is to make the error messages
look the same, i.e.:

run:1: command not found: a-non-existent-command

instead of (eval):1: command .... I've checked that zunit's tests pass normally
with this change.